refactor(hooks): share git push target resolution via _lib
push-remote-ref-verify held the only parser that answers "after this push,
which remote branch is supposed to have moved?". A second PostToolUse hook
needs the identical answer, and that parser is where all the difficulty
lives — `git -C`, value-taking flags, `local:remote` refspecs, bare pushes
resolved through @{upstream}, detached HEAD, branch deletion.

_lib is the established sharing surface here (53 hooks import from it);
cross-hook imports have one precedent. Behavior is unchanged — the hook's
existing 17-case suite passes untouched.

Confidence: high
