+++ hooks/advisory-nudge/push-remote-ref-verify/impl.py
@@ -9,6 +9,9 @@ branch never receives the commit even though `git push` prints
 and main goes red. The `* [new branch]` line on an already-pushed branch is the
 tell, but it is easy to miss in the moment.
 
+The `git push` argv parsing lives in `_lib/_git_push_target.py` (issue #1039)
+because a second PostToolUse hook needs the same answer.
+
 This hook runs AFTER a `git push` and cross-checks the remote tip against the
 SHA that was supposed to be pushed:
 
@@ -36,176 +39,20 @@ from __future__ import annotations
 import json
 import os
 import re
-import subprocess
 import sys
 from pathlib import Path as _Path
 
 sys.path.insert(0, str(_Path(__file__).resolve().parent.parent.parent / "_lib"))
-from _hook_runtime import fail_open  # type: ignore[import-not-found]  # noqa: E402
-from _hook_utils import (  # type: ignore[import-not-found]  # noqa: E402
-    iter_command_starts,
-    safe_tokenize,
-    strip_prefix,
+from _git_push_target import (  # type: ignore[import-not-found]  # noqa: E402
+    git as _git,
 )
+from _git_push_target import resolve_push_target  # type: ignore[import-not-found]  # noqa: E402
+from _hook_runtime import fail_open  # type: ignore[import-not-found]  # noqa: E402
 
 _BYPASS_ENV = "PRAXIS_PUSH_VERIFY_BYPASS"
 _STRICT_ENV = "PRAXIS_PUSH_VERIFY_STRICT"
-_GIT_TIMEOUT_SEC = 10
-
-# git push flags we cannot reason about safely -> skip the whole push.
-_SKIP_FLAGS = frozenset({
-    "--dry-run", "-n", "--delete", "-d", "--tags", "--all",
-    "--mirror", "--prune",
-})
-# git push flags that consume the following token as a value.
-_VALUE_FLAGS = frozenset({"-o", "--push-option", "--exec", "--receive-pack", "--repo"})
 
 _SHA_RE = re.compile(r"^[0-9a-f]{7,40}$")
-
-
-# ---------------------------------------------------------------------------
-# git helpers (fail-open)
-# ---------------------------------------------------------------------------
-
-
-def _git(cwd: str, args: list[str], timeout: int = _GIT_TIMEOUT_SEC):
-    """Run `git -C <cwd> <args>`. Return (returncode|None, stdout). Never raises."""
-    try:
-        r = subprocess.run(
-            ["git", "-C", cwd, *args],
-            capture_output=True,
-            text=True,
-            timeout=timeout,
-        )
-        return r.returncode, (r.stdout or "")
-    except (OSError, subprocess.SubprocessError):
-        return None, ""
-
-
-# ---------------------------------------------------------------------------
-# command parsing
-# ---------------------------------------------------------------------------
-
-
-def _find_push_argv(command: str):
-    """Return (push_args, git_C_dir|None) for the first `git push` segment, else None."""
-    try:
-        tokens = safe_tokenize(command)
-    except Exception:
-        return None
-    for argv in iter_command_starts(tokens):
-        argv = strip_prefix(argv)
-        if not argv or argv[0] != "git":
-            continue
-        # Walk global git options (-C <dir>, -c <kv>, --git-dir=...) to the subcommand.
-        i = 1
-        git_c = None
-        while i < len(argv):
-            tok = argv[i]
-            if tok == "-C" and i + 1 < len(argv):
-                git_c = argv[i + 1]
-                i += 2
-                continue
-            if tok == "-c" and i + 1 < len(argv):
-                i += 2
-                continue
-            if tok.startswith("-"):
-                i += 1
-                continue
-            break
-        if i < len(argv) and argv[i] == "push":
-            return argv[i + 1:], git_c
-    return None
-
-
-def _parse_push(push_args: list[str]):
-    """Extract {remote, refspec} from push argv, or None to skip."""
-    positionals: list[str] = []
-    i = 0
-    while i < len(push_args):
-        tok = push_args[i]
-        if tok in _SKIP_FLAGS:
-            return None
-        if tok in ("--force", "-f", "--force-with-lease") or tok.startswith("--force-with-lease="):
-            i += 1
-            continue
-        if tok in _VALUE_FLAGS:
-            i += 2  # consume the value token too
-            continue
-        if tok.startswith("--") and "=" in tok:
-            key = tok.split("=", 1)[0]
-            if key in _SKIP_FLAGS:
-                return None
-            if key in _VALUE_FLAGS:
-                i += 1  # value already inline after `=`; no next token consumed
-                continue
-            i += 1
-            continue
-        if tok.startswith("-"):
-            i += 1  # boolean short/long flag (e.g. -u, --set-upstream, --atomic)
-            continue
-        positionals.append(tok)
-        i += 1
-    if len(positionals) > 2:
-        return None  # multiple refspecs — too ambiguous to verify safely
-    remote = positionals[0] if positionals else None
-    refspec = positionals[1] if len(positionals) >= 2 else None
-    return {"remote": remote, "refspec": refspec}
-
-
-def _resolve_targets(cwd: str, parsed: dict):
-    """Resolve {remote, remote_branch, local_ref} for the push, or None to skip."""
-    remote = parsed["remote"]
-    refspec = parsed["refspec"]
-
-    def _current_branch():
-        rc, cur = _git(cwd, ["rev-parse", "--abbrev-ref", "HEAD"])
-        cur = cur.strip()
-        if rc != 0 or not cur or cur == "HEAD":
-            return None  # detached HEAD
-        return cur
-
-    if refspec is None:
-        cur = _current_branch()
-        if cur is None:
-            return None
-        if remote is None:
-            rc, up = _git(
-                cwd, ["rev-parse", "--abbrev-ref", "--symbolic-full-name", "@{upstream}"]
-            )
-            up = up.strip()
-            if rc != 0 or "/" not in up:
-                return None  # no upstream configured — cannot know the target
-            remote, remote_branch = up.split("/", 1)
-        else:
-            remote_branch = cur
-        local_ref = "HEAD"
-    else:
-        if refspec.startswith(":"):
-            return None  # branch deletion (`git push origin :branch`)
-        if refspec.startswith("+"):
-            refspec = refspec[1:]
-        if remote is None:
-            remote = "origin"
-        if ":" in refspec:
-            local_part, remote_branch = refspec.split(":", 1)
-            local_ref = local_part or "HEAD"
-        elif refspec == "HEAD":
-            cur = _current_branch()
-            if cur is None:
-                return None
-            remote_branch = cur
-            local_ref = "HEAD"
-        else:
-            local_ref = refspec
-            remote_branch = refspec
-        remote_branch = remote_branch.strip()
-        if remote_branch.startswith("refs/heads/"):
-            remote_branch = remote_branch[len("refs/heads/"):]
-
-    if not remote or not remote_branch:
-        return None
-    return {"remote": remote, "remote_branch": remote_branch, "local_ref": local_ref}
 
 
 # ---------------------------------------------------------------------------
@@ -291,18 +138,10 @@ def main() -> int:
         if tool_response.get("interrupted") is True or tool_response.get("isError") is True:
             return 0
 
-    found = _find_push_argv(command)
-    if not found:
-        return 0
-    push_args, git_c = found
-    parsed = _parse_push(push_args)
-    if parsed is None:
-        return 0
-
-    cwd = git_c or payload.get("cwd") or os.getcwd()
-    targets = _resolve_targets(cwd, parsed)
+    targets = resolve_push_target(command, payload.get("cwd") or os.getcwd())
     if targets is None:
         return 0
+    cwd = targets["cwd"]
 
     expected_rc, expected_sha = _git(cwd, ["rev-parse", "--verify", targets["local_ref"]])
     expected_sha = expected_sha.strip()
